Move specific check methods to separate classes
