+++ network/datadog_checks/network/network.py
@@ -15,13 +15,12 @@ from collections import defaultdict
 import psutil
 from six import PY3, iteritems, itervalues
 
-from datadog_checks.base import AgentCheck, ConfigurationError, is_affirmative
-from datadog_checks.base.utils.common import pattern_filter
+from datadog_checks.base import AgentCheck, ConfigurationError
 from datadog_checks.base.utils.platform import Platform
 from datadog_checks.base.utils.subprocess_output import SubprocessOutputEmptyError, get_subprocess_output
 
 from . import ethtool
-from .const import BSD_TCP_METRICS, ENA_METRIC_NAMES, ENA_METRIC_PREFIX, SOLARIS_TCP_METRICS
+from .const import ENA_METRIC_NAMES, ENA_METRIC_PREFIX
 
 try:
     import datadog_agent
@@ -348,277 +347,6 @@ class Network(AgentCheck):
 
         return True
 
-    def _check_linux(self, instance):
-        """
-        _check_linux can be run inside a container and still collects the network metrics from the host
-        For that procfs_path can be set to something like "/host/proc"
-        When a custom procfs_path is set, the collect_connection_state option is ignored
-        """
-        proc_location = datadog_agent.get_config('procfs_path')
-        if not proc_location:
-            proc_location = '/proc'
-        proc_location = proc_location.rstrip('/')
-        custom_tags = instance.get('tags', [])
-
-        self._get_iface_sys_metrics(custom_tags)
-        net_proc_base_location = self._get_net_proc_base_location(proc_location)
-
-        if self._is_collect_cx_state_runnable(net_proc_base_location):
-            try:
-                self.log.debug("Using `ss` to collect connection state")
-                # Try using `ss` for increased performance over `netstat`
-                ss_env = {"PROC_ROOT": net_proc_base_location}
-
-                # By providing the environment variables in ss_env, the PATH will be overridden. In CentOS,
-                # datadog-agent PATH is "/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin", while sh PATH
-                # will be '/usr/local/bin:/usr/bin'. In CentOS, ss is located in /sbin and /usr/sbin, not
-                # in the sh PATH, which will result in network metric collection failure.
-                #
-                # The line below will set sh PATH explicitly as the datadog-agent PATH to fix that issue.
-                if "PATH" in os.environ:
-                    ss_env["PATH"] = os.environ["PATH"]
-
-                metrics = self._get_metrics()
-                for ip_version in ['4', '6']:
-                    # Call `ss` for each IP version because there's no built-in way of distinguishing
-                    # between the IP versions in the output
-                    # Also calls `ss` for each protocol, because on some systems (e.g. Ubuntu 14.04), there is a
-                    # bug that print `tcp` even if it's `udp`
-                    # The `-H` flag isn't available on old versions of `ss`.
-                    cmd = "ss --numeric --tcp --all --ipv{} | cut -d ' ' -f 1 | sort | uniq -c".format(ip_version)
-                    output, _, _ = get_subprocess_output(["sh", "-c", cmd], self.log, env=ss_env)
-
-                    # 7624 CLOSE-WAIT
-                    #   72 ESTAB
-                    #    9 LISTEN
-                    #    1 State
-                    #   37 TIME-WAIT
-                    lines = output.splitlines()
-
-                    self._parse_short_state_lines(lines, metrics, self.tcp_states['ss'], ip_version=ip_version)
-
-                    cmd = "ss --numeric --udp --all --ipv{} | wc -l".format(ip_version)
-                    output, _, _ = get_subprocess_output(["sh", "-c", cmd], self.log, env=ss_env)
-                    metric = self.cx_state_gauge[('udp{}'.format(ip_version), 'connections')]
-                    metrics[metric] = int(output) - 1  # Remove header
-
-                    if self._collect_cx_queues:
-                        cmd = "ss --numeric --tcp --all --ipv{}".format(ip_version)
-                        output, _, _ = get_subprocess_output(["sh", "-c", cmd], self.log, env=ss_env)
-                        for (state, recvq, sendq) in self._parse_queues("ss", output):
-                            self.histogram('system.net.tcp.recv_q', recvq, custom_tags + ["state:" + state])
-                            self.histogram('system.net.tcp.send_q', sendq, custom_tags + ["state:" + state])
-
-                for metric, value in iteritems(metrics):
-                    self.gauge(metric, value, tags=custom_tags)
-
-            except OSError as e:
-                self.log.info("`ss` invocation failed: %s. Using `netstat` as a fallback", str(e))
-                output, _, _ = get_subprocess_output(["netstat", "-n", "-u", "-t", "-a"], self.log)
-                lines = output.splitlines()
-                # Active Internet connections (w/o servers)
-                # Proto Recv-Q Send-Q Local Address           Foreign Address         State
-                # tcp        0      0 46.105.75.4:80          79.220.227.193:2032     SYN_RECV
-                # tcp        0      0 46.105.75.4:143         90.56.111.177:56867     ESTABLISHED
-                # tcp        0      0 46.105.75.4:50468       107.20.207.175:443      TIME_WAIT
-                # tcp6       0      0 46.105.75.4:80          93.15.237.188:58038     FIN_WAIT2
-                # tcp6       0      0 46.105.75.4:80          79.220.227.193:2029     ESTABLISHED
-                # udp        0      0 0.0.0.0:123             0.0.0.0:*
-                # udp6       0      0 :::41458                :::*
-
-                metrics = self._parse_linux_cx_state(lines[2:], self.tcp_states['netstat'], 5)
-                for metric, value in iteritems(metrics):
-                    self.gauge(metric, value, tags=custom_tags)
-
-                if self._collect_cx_queues:
-                    for (state, recvq, sendq) in self._parse_queues("netstat", output):
-                        self.histogram('system.net.tcp.recv_q', recvq, custom_tags + ["state:" + state])
-                        self.histogram('system.net.tcp.send_q', sendq, custom_tags + ["state:" + state])
-
-            except SubprocessOutputEmptyError:
-                self.log.exception("Error collecting connection states.")
-
-        proc_dev_path = "{}/net/dev".format(net_proc_base_location)
-        try:
-            with open(proc_dev_path, 'r') as proc:
-                lines = proc.readlines()
-        except IOError:
-            # On Openshift, /proc/net/snmp is only readable by root
-            self.log.debug("Unable to read %s.", proc_dev_path)
-            lines = []
-
-        # Inter-|   Receive                                                 |  Transmit
-        #  face |bytes     packets errs drop fifo frame compressed multicast|bytes       packets errs drop fifo colls carrier compressed # noqa: E501
-        #     lo:45890956   112797   0    0    0     0          0         0    45890956   112797    0    0    0     0       0          0 # noqa: E501
-        #   eth0:631947052 1042233   0   19    0   184          0      1206  1208625538  1320529    0    0    0     0       0          0 # noqa: E501
-        #   eth1:       0        0   0    0    0     0          0         0           0        0    0    0    0     0       0          0 # noqa: E501
-        for line in lines[2:]:
-            cols = line.split(':', 1)
-            x = cols[1].split()
-            # Filter inactive interfaces
-            if self._parse_value(x[0]) or self._parse_value(x[8]):
-                iface = cols[0].strip()
-                metrics = {
-                    'bytes_rcvd': self._parse_value(x[0]),
-                    'bytes_sent': self._parse_value(x[8]),
-                    'packets_in.count': self._parse_value(x[1]),
-                    'packets_in.drop': self._parse_value(x[3]),
-                    'packets_in.error': self._parse_value(x[2]) + self._parse_value(x[3]),
-                    'packets_out.count': self._parse_value(x[9]),
-                    'packets_out.drop': self._parse_value(x[11]),
-                    'packets_out.error': self._parse_value(x[10]) + self._parse_value(x[11]),
-                }
-                self._submit_devicemetrics(iface, metrics, custom_tags)
-                self._handle_ethtool_stats(iface, custom_tags)
-
-        netstat_data = {}
-        for f in ['netstat', 'snmp']:
-            proc_data_path = "{}/net/{}".format(net_proc_base_location, f)
-            try:
-                with open(proc_data_path, 'r') as netstat:
-                    while True:
-                        n_header = netstat.readline()
-                        if not n_header:
-                            break  # No more? Abort!
-                        n_data = netstat.readline()
-
-                        h_parts = n_header.strip().split(' ')
-                        h_values = n_data.strip().split(' ')
-                        ns_category = h_parts[0][:-1]
-                        netstat_data[ns_category] = {}
-                        # Turn the data into a dictionary
-                        for idx, hpart in enumerate(h_parts[1:]):
-                            netstat_data[ns_category][hpart] = h_values[idx + 1]
-            except IOError:
-                # On Openshift, /proc/net/snmp is only readable by root
-                self.log.debug("Unable to read %s.", proc_data_path)
-
-        nstat_metrics_names = {
-            'Ip': {
-                'InReceives': 'system.net.ip.in_receives',
-                'InHdrErrors': 'system.net.ip.in_header_errors',
-                'InAddrErrors': 'system.net.ip.in_addr_errors',
-                'InUnknownProtos': 'system.net.ip.in_unknown_protos',
-                'InDiscards': 'system.net.ip.in_discards',
-                'InDelivers': 'system.net.ip.in_delivers',
-                'OutRequests': 'system.net.ip.out_requests',
-                'OutDiscards': 'system.net.ip.out_discards',
-                'OutNoRoutes': 'system.net.ip.out_no_routes',
-                'ForwDatagrams': 'system.net.ip.forwarded_datagrams',
-                'ReasmTimeout': 'system.net.ip.reassembly_timeouts',
-                'ReasmReqds': 'system.net.ip.reassembly_requests',
-                'ReasmOKs': 'system.net.ip.reassembly_oks',
-                'ReasmFails': 'system.net.ip.reassembly_fails',
-                'FragOKs': 'system.net.ip.fragmentation_oks',
-                'FragFails': 'system.net.ip.fragmentation_fails',
-                'FragCreates': 'system.net.ip.fragmentation_creates',
-            },
-            'IpExt': {
-                'InNoRoutes': 'system.net.ip.in_no_routes',
-                'InTruncatedPkts': 'system.net.ip.in_truncated_pkts',
-                'InCsumErrors': 'system.net.ip.in_csum_errors',
-                'ReasmOverlaps': 'system.net.ip.reassembly_overlaps',
-            },
-            'Tcp': {
-                'RetransSegs': 'system.net.tcp.retrans_segs',
-                'InSegs': 'system.net.tcp.in_segs',
-                'OutSegs': 'system.net.tcp.out_segs',
-                'ActiveOpens': 'system.net.tcp.active_opens',
-                'PassiveOpens': 'system.net.tcp.passive_opens',
-                'AttemptFails': 'system.net.tcp.attempt_fails',
-                'EstabResets': 'system.net.tcp.established_resets',
-                'InErrs': 'system.net.tcp.in_errors',
-                'OutRsts': 'system.net.tcp.out_resets',
-                'InCsumErrors': 'system.net.tcp.in_csum_errors',
-            },
-            'TcpExt': {
-                'ListenOverflows': 'system.net.tcp.listen_overflows',
-                'ListenDrops': 'system.net.tcp.listen_drops',
-                'TCPBacklogDrop': 'system.net.tcp.backlog_drops',
-                'TCPRetransFail': 'system.net.tcp.failed_retransmits',
-                'IPReversePathFilter': 'system.net.ip.reverse_path_filter',
-                'PruneCalled': 'system.net.tcp.prune_called',
-                'RcvPruned': 'system.net.tcp.prune_rcv_drops',
-                'OfoPruned': 'system.net.tcp.prune_ofo_called',
-                'PAWSActive': 'system.net.tcp.paws_connection_drops',
-                'PAWSEstab': 'system.net.tcp.paws_established_drops',
-                'SyncookiesSent': 'system.net.tcp.syn_cookies_sent',
-                'SyncookiesRecv': 'system.net.tcp.syn_cookies_recv',
-                'SyncookiesFailed': 'system.net.tcp.syn_cookies_failed',
-                'TCPAbortOnTimeout': 'system.net.tcp.abort_on_timeout',
-                'TCPSynRetrans': 'system.net.tcp.syn_retrans',
-                'TCPFromZeroWindowAdv': 'system.net.tcp.from_zero_window',
-                'TCPToZeroWindowAdv': 'system.net.tcp.to_zero_window',
-                'TWRecycled': 'system.net.tcp.tw_reused',
-            },
-            'Udp': {
-                'InDatagrams': 'system.net.udp.in_datagrams',
-                'NoPorts': 'system.net.udp.no_ports',
-                'InErrors': 'system.net.udp.in_errors',
-                'OutDatagrams': 'system.net.udp.out_datagrams',
-                'RcvbufErrors': 'system.net.udp.rcv_buf_errors',
-                'SndbufErrors': 'system.net.udp.snd_buf_errors',
-                'InCsumErrors': 'system.net.udp.in_csum_errors',
-            },
-        }
-        nstat_metrics_gauge_names = {
-            'Tcp': {
-                'CurrEstab': 'system.net.tcp.current_established',
-            },
-        }
-
-        for k in nstat_metrics_names:
-            for met in nstat_metrics_names[k]:
-                if met in netstat_data.get(k, {}):
-                    self._submit_netmetric(
-                        nstat_metrics_names[k][met], self._parse_value(netstat_data[k][met]), tags=custom_tags
-                    )
-
-        for k in nstat_metrics_gauge_names:
-            for met in nstat_metrics_gauge_names[k]:
-                if met in netstat_data.get(k, {}):
-                    self._submit_netmetric_gauge(
-                        nstat_metrics_gauge_names[k][met], self._parse_value(netstat_data[k][met]), tags=custom_tags
-                    )
-
-        # Get the conntrack -S information
-        conntrack_path = instance.get('conntrack_path')
-        use_sudo_conntrack = is_affirmative(instance.get('use_sudo_conntrack', True))
-        if conntrack_path is not None:
-            self._add_conntrack_stats_metrics(conntrack_path, use_sudo_conntrack, custom_tags)
-
-        # Get the rest of the metric by reading the files. Metrics available since kernel 3.6
-        conntrack_files_location = os.path.join(proc_location, 'sys', 'net', 'netfilter')
-        # By default, only max and count are reported. However if the blacklist is set,
-        # the whitelist is losing its default value
-        blacklisted_files = instance.get('blacklist_conntrack_metrics')
-        whitelisted_files = instance.get('whitelist_conntrack_metrics')
-        if blacklisted_files is None and whitelisted_files is None:
-            whitelisted_files = ['max', 'count']
-
-        available_files = []
-
-        # Get the metrics to read
-        try:
-            for metric_file in os.listdir(conntrack_files_location):
-                if (
-                    os.path.isfile(os.path.join(conntrack_files_location, metric_file))
-                    and 'nf_conntrack_' in metric_file
-                ):
-                    available_files.append(metric_file[len('nf_conntrack_') :])
-        except Exception as e:
-            self.log.debug("Unable to list the files in %s. %s", conntrack_files_location, e)
-
-        filtered_available_files = pattern_filter(
-            available_files, whitelist=whitelisted_files, blacklist=blacklisted_files
-        )
-
-        for metric_name in filtered_available_files:
-            metric_file_location = os.path.join(conntrack_files_location, 'nf_conntrack_{}'.format(metric_name))
-            value = self._read_int_file(metric_file_location)
-            if value is not None:
-                self.gauge('system.net.conntrack.{}'.format(metric_name), value, tags=custom_tags)
-
     @staticmethod
     def _get_net_proc_base_location(proc_location):
         if Platform.is_containerized() and proc_location != "/proc":
@@ -727,254 +455,6 @@ class Network(AgentCheck):
                 metrics[metric] += 1
 
         return metrics
-
-    def _check_bsd(self, instance):
-        netstat_flags = ['-i', '-b']
-
-        custom_tags = instance.get('tags', [])
-
-        # FreeBSD's netstat truncates device names unless you pass '-W'
-        if Platform.is_freebsd():
-            netstat_flags.append('-W')
-
-        try:
-            output, _, _ = get_subprocess_output(["netstat"] + netstat_flags, self.log)
-            lines = output.splitlines()
-            # Name  Mtu   Network       Address            Ipkts Ierrs     Ibytes    Opkts Oerrs     Obytes  Coll
-            # lo0   16384 <Link#1>                        318258     0  428252203   318258     0  428252203     0
-            # lo0   16384 localhost   fe80:1::1           318258     -  428252203   318258     -  428252203     -
-            # lo0   16384 127           localhost         318258     -  428252203   318258     -  428252203     -
-            # lo0   16384 localhost   ::1                 318258     -  428252203   318258     -  428252203     -
-            # gif0* 1280  <Link#2>                             0     0          0        0     0          0     0
-            # stf0* 1280  <Link#3>                             0     0          0        0     0          0     0
-            # en0   1500  <Link#4>    04:0c:ce:db:4e:fa 20801309     0 13835457425 15149389     0 11508790198     0
-            # en0   1500  seneca.loca fe80:4::60c:ceff: 20801309     - 13835457425 15149389     - 11508790198     -
-            # en0   1500  2001:470:1f 2001:470:1f07:11d 20801309     - 13835457425 15149389     - 11508790198     -
-            # en0   1500  2001:470:1f 2001:470:1f07:11d 20801309     - 13835457425 15149389     - 11508790198     -
-            # en0   1500  192.168.1     192.168.1.63    20801309     - 13835457425 15149389     - 11508790198     -
-            # en0   1500  2001:470:1f 2001:470:1f07:11d 20801309     - 13835457425 15149389     - 11508790198     -
-            # p2p0  2304  <Link#5>    06:0c:ce:db:4e:fa        0     0          0        0     0          0     0
-            # ham0  1404  <Link#6>    7a:79:05:4d:bf:f5    30100     0    6815204    18742     0    8494811     0
-            # ham0  1404  5             5.77.191.245       30100     -    6815204    18742     -    8494811     -
-            # ham0  1404  seneca.loca fe80:6::7879:5ff:    30100     -    6815204    18742     -    8494811     -
-            # ham0  1404  2620:9b::54 2620:9b::54d:bff5    30100     -    6815204    18742     -    8494811     -
-
-            headers = lines[0].split()
-
-            # Given the irregular structure of the table above, better to parse from the end of each line
-            # Verify headers first
-            #          -7       -6       -5        -4       -3       -2        -1
-            for h in ("Ipkts", "Ierrs", "Ibytes", "Opkts", "Oerrs", "Obytes", "Coll"):
-                if h not in headers:
-                    self.log.error("%s not found in %s; cannot parse", h, headers)
-                    return False
-
-            current = None
-            for l in lines[1:]:
-                # Another header row, abort now, this is IPv6 land
-                if "Name" in l:
-                    break
-
-                x = l.split()
-                if len(x) == 0:
-                    break
-
-                iface = x[0]
-                if iface.endswith("*"):
-                    iface = iface[:-1]
-                if iface == current:
-                    # skip multiple lines of same interface
-                    continue
-                else:
-                    current = iface
-
-                # Filter inactive interfaces
-                if self._parse_value(x[-5]) or self._parse_value(x[-2]):
-                    iface = current
-                    metrics = {
-                        'bytes_rcvd': self._parse_value(x[-5]),
-                        'bytes_sent': self._parse_value(x[-2]),
-                        'packets_in.count': self._parse_value(x[-7]),
-                        'packets_in.error': self._parse_value(x[-6]),
-                        'packets_out.count': self._parse_value(x[-4]),
-                        'packets_out.error': self._parse_value(x[-3]),
-                    }
-                    self._submit_devicemetrics(iface, metrics, custom_tags)
-        except SubprocessOutputEmptyError:
-            self.log.exception("Error collecting connection stats.")
-
-        try:
-            netstat, _, _ = get_subprocess_output(["netstat", "-s", "-p" "tcp"], self.log)
-            # 3651535 packets sent
-            #         972097 data packets (615753248 bytes)
-            #         5009 data packets (2832232 bytes) retransmitted
-            #         0 resends initiated by MTU discovery
-            #         2086952 ack-only packets (471 delayed)
-            #         0 URG only packets
-            #         0 window probe packets
-            #         310851 window update packets
-            #         336829 control packets
-            #         0 data packets sent after flow control
-            #         3058232 checksummed in software
-            #         3058232 segments (571218834 bytes) over IPv4
-            #         0 segments (0 bytes) over IPv6
-            # 4807551 packets received
-            #         1143534 acks (for 616095538 bytes)
-            #         165400 duplicate acks
-            #         ...
-
-            self._submit_regexed_values(netstat, BSD_TCP_METRICS, custom_tags)
-        except SubprocessOutputEmptyError:
-            self.log.exception("Error collecting TCP stats.")
-
-        proc_location = self.agentConfig.get('procfs_path', '/proc').rstrip('/')
-
-        net_proc_base_location = self._get_net_proc_base_location(proc_location)
-
-        if self._is_collect_cx_state_runnable(net_proc_base_location):
-            try:
-                self.log.debug("Using `netstat` to collect connection state")
-                output_TCP, _, _ = get_subprocess_output(["netstat", "-n", "-a", "-p", "tcp"], self.log)
-                output_UDP, _, _ = get_subprocess_output(["netstat", "-n", "-a", "-p", "udp"], self.log)
-                lines = output_TCP.splitlines() + output_UDP.splitlines()
-                # Active Internet connections (w/o servers)
-                # Proto Recv-Q Send-Q Local Address           Foreign Address         State
-                # tcp        0      0 46.105.75.4:80          79.220.227.193:2032     SYN_RECV
-                # tcp        0      0 46.105.75.4:143         90.56.111.177:56867     ESTABLISHED
-                # tcp        0      0 46.105.75.4:50468       107.20.207.175:443      TIME_WAIT
-                # tcp6       0      0 46.105.75.4:80          93.15.237.188:58038     FIN_WAIT2
-                # tcp6       0      0 46.105.75.4:80          79.220.227.193:2029     ESTABLISHED
-                # udp        0      0 0.0.0.0:123             0.0.0.0:*
-                # udp6       0      0 :::41458                :::*
-
-                metrics = self._parse_linux_cx_state(lines[2:], self.tcp_states['netstat'], 5)
-                for metric, value in iteritems(metrics):
-                    self.gauge(metric, value, tags=custom_tags)
-            except SubprocessOutputEmptyError:
-                self.log.exception("Error collecting connection states.")
-
-    def _check_solaris(self, instance):
-        # Can't get bytes sent and received via netstat
-        # Default to kstat -p link:0:
-        custom_tags = instance.get('tags', [])
-        try:
-            netstat, _, _ = get_subprocess_output(["kstat", "-p", "link:0:"], self.log)
-            metrics_by_interface = self._parse_solaris_netstat(netstat)
-            for interface, metrics in iteritems(metrics_by_interface):
-                self._submit_devicemetrics(interface, metrics, custom_tags)
-        except SubprocessOutputEmptyError:
-            self.log.exception("Error collecting kstat stats.")
-
-        try:
-            netstat, _, _ = get_subprocess_output(["netstat", "-s", "-P" "tcp"], self.log)
-            # TCP: tcpRtoAlgorithm=     4 tcpRtoMin           =   200
-            # tcpRtoMax           = 60000 tcpMaxConn          =    -1
-            # tcpActiveOpens      =    57 tcpPassiveOpens     =    50
-            # tcpAttemptFails     =     1 tcpEstabResets      =     0
-            # tcpCurrEstab        =     0 tcpOutSegs          =   254
-            # tcpOutDataSegs      =   995 tcpOutDataBytes     =1216733
-            # tcpRetransSegs      =     0 tcpRetransBytes     =     0
-            # tcpOutAck           =   185 tcpOutAckDelayed    =     4
-            # ...
-            self._submit_regexed_values(netstat, SOLARIS_TCP_METRICS, custom_tags)
-        except SubprocessOutputEmptyError:
-            self.log.exception("Error collecting TCP stats.")
-
-    def _parse_solaris_netstat(self, netstat_output):
-        """
-        Return a mapping of network metrics by interface. For example:
-            { interface:
-                {'bytes_sent': 0,
-                  'bytes_rcvd': 0,
-                  'bytes_rcvd': 0,
-                  ...
-                }
-            }
-        """
-        # Here's an example of the netstat output:
-        #
-        # link:0:net0:brdcstrcv   527336
-        # link:0:net0:brdcstxmt   1595
-        # link:0:net0:class       net
-        # link:0:net0:collisions  0
-        # link:0:net0:crtime      16359935.2637943
-        # link:0:net0:ierrors     0
-        # link:0:net0:ifspeed     10000000000
-        # link:0:net0:ipackets    682834
-        # link:0:net0:ipackets64  682834
-        # link:0:net0:link_duplex 0
-        # link:0:net0:link_state  1
-        # link:0:net0:multircv    0
-        # link:0:net0:multixmt    1595
-        # link:0:net0:norcvbuf    0
-        # link:0:net0:noxmtbuf    0
-        # link:0:net0:obytes      12820668
-        # link:0:net0:obytes64    12820668
-        # link:0:net0:oerrors     0
-        # link:0:net0:opackets    105445
-        # link:0:net0:opackets64  105445
-        # link:0:net0:rbytes      113983614
-        # link:0:net0:rbytes64    113983614
-        # link:0:net0:snaptime    16834735.1607669
-        # link:0:net0:unknowns    0
-        # link:0:net0:zonename    53aa9b7e-48ba-4152-a52b-a6368c3d9e7c
-        # link:0:net1:brdcstrcv   4947620
-        # link:0:net1:brdcstxmt   1594
-        # link:0:net1:class       net
-        # link:0:net1:collisions  0
-        # link:0:net1:crtime      16359935.2839167
-        # link:0:net1:ierrors     0
-        # link:0:net1:ifspeed     10000000000
-        # link:0:net1:ipackets    4947620
-        # link:0:net1:ipackets64  4947620
-        # link:0:net1:link_duplex 0
-        # link:0:net1:link_state  1
-        # link:0:net1:multircv    0
-        # link:0:net1:multixmt    1594
-        # link:0:net1:norcvbuf    0
-        # link:0:net1:noxmtbuf    0
-        # link:0:net1:obytes      73324
-        # link:0:net1:obytes64    73324
-        # link:0:net1:oerrors     0
-        # link:0:net1:opackets    1594
-        # link:0:net1:opackets64  1594
-        # link:0:net1:rbytes      304384894
-        # link:0:net1:rbytes64    304384894
-        # link:0:net1:snaptime    16834735.1613302
-        # link:0:net1:unknowns    0
-        # link:0:net1:zonename    53aa9b7e-48ba-4152-a52b-a6368c3d9e7c
-
-        # A mapping of solaris names -> datadog names
-        metric_by_solaris_name = {
-            'rbytes64': 'bytes_rcvd',
-            'obytes64': 'bytes_sent',
-            'ipackets64': 'packets_in.count',
-            'ierrors': 'packets_in.error',
-            'opackets64': 'packets_out.count',
-            'oerrors': 'packets_out.error',
-        }
-
-        lines = [l for l in netstat_output.splitlines() if len(l) > 0]
-
-        metrics_by_interface = {}
-
-        for l in lines:
-            # Parse the metric & interface.
-            cols = l.split()
-            link, n, iface, name = cols[0].split(":")
-            assert link == "link"
-
-            # Get the datadog metric name.
-            ddname = metric_by_solaris_name.get(name, None)
-            if ddname is None:
-                continue
-
-            # Add it to this interface's list of metrics.
-            metrics = metrics_by_interface.get(iface, {})
-            metrics[ddname] = self._parse_value(cols[1])
-            metrics_by_interface[iface] = metrics
-
-        return metrics_by_interface
 
     def _check_psutil(self, instance):
         """
